+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Slightly changed the design of the review summary
+
 ## [2.0.0] - 2021-02-23
 
 ### Added