+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Legacy reviews display store block
+
+### Fixed
+- Duplicate insertion of `power-reviews` script in the DOM
+- Sometimes the Q&A block wasn't being rendered at all
+
 ## [2.0.0] - 2021-02-23
 
 ### Added